% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btree_paths.R
\name{btree_paths}
\alias{btree_paths}
\title{btree Paths}
\usage{
btree_paths(btree)
}
\arguments{
\item{btree}{A btree}
}
\description{
Given a btree, produce a vector of strings indicating the directed steps to take starting at the root to get to each node
}
\details{
Returns a character vector whose values correspond to each node in the btree
}
\examples{
library(data.table)
mytree <- data.table(NodeId=c(1, 2, 3), ParentNodeId=c(NA, 1, 1))
btree_paths(mytree)
}
