% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_categorical_splits.R
\name{describe_categorical_splits}
\alias{describe_categorical_splits}
\title{Describe Categorical Splits}
\usage{
describe_categorical_splits(rules, catVars)
}
\arguments{
\item{rules}{fill me in}
}
\description{
Helper function for make_dtree()
}
\details{
fill me in
}
