% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree_node_conditions.R
\name{dtree_node_conditions}
\alias{dtree_node_conditions}
\title{dtree Node Conditions}
\usage{
dtree_node_conditions(dtree)
}
\arguments{
\item{dtree}{A dtree}
}
\description{
For each node in a dtree, get the logical conditions that send samples to the node
}
\details{
Returns a data.table with columns {NodeId, Condition}
}
\examples{
library(data.table)

dtree1 <- make_dtree(nodeIds=c(1), leftChildIds=c(2), rightChildIds=c(3), splitVars=c("foo"), splitVals=c(3.5))
dtree_node_conditions(dtree1)
}
