% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_one_root.R
\name{has_one_root}
\alias{has_one_root}
\title{Has One Root}
\usage{
has_one_root(btree)
}
\arguments{
\item{btree}{A data.table object with field \code{ParentNodeID}}
}
\description{
Helper method for validating a data.table object as a btree by checking for one root.
}
\details{
Returns \code{TRUE} if field \code{ParentNodeID} has exactly 1 NA value.
}
\examples{
library(data.table)
mytree <- data.table(NodeID=c(1, 2, 3), ParentNodeID=c(NA, 1, 1))
has_one_root(mytree)
}

