% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_two_parents.R
\name{has_two_children}
\alias{has_two_children}
\title{Has Two Children}
\usage{
has_two_children(btree)
}
\arguments{
\item{btree}{A data.table object with fields \code{NodeId} and \code{ParentNodeId}}
}
\description{
Helper method for validating a data.table object as a btree by checking for two children per node, excluding leaves.
}
\details{
Returns \code{TRUE} or \code{FALSE}
\code{NodeId}
}
\examples{
library(data.table)
mytree <- data.table(NodeId=c(1, 2, 3), ParentNodeId=c(NA, 1, 1))
has_two_children(mytree)
}
