% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_two_parents.R
\name{has_two_parents}
\alias{has_two_parents}
\title{Has Two Parents}
\usage{
has_two_parents(btree)
}
\arguments{
\item{btree}{A data.table object with fields \code{NodeID} and \code{ParentNodeID}}
}
\description{
Helper method for validating a data.table object as a btree by checking for two parents per node, excluding the root node.
}
\details{
Returns \code{TRUE} if every unique \code{ParentNodeID} besides NA (the root) has exactly 2 matching values in field
\code{NodeID}
}
\examples{
library(data.table)
mytree <- data.table(NodeID=c(1, 2, 3), ParentNodeID=c(NA, 1, 1))
has_two_parents(mytree)
}

