% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_btree.R
\name{make_btree}
\alias{make_btree}
\title{Make btree}
\usage{
make_btree(nodeIDs, parentNodeIDs = NULL, leftChildIDs = NULL,
  rightChildIDs = NULL)
}
\arguments{
\item{nodeIDs}{A vector of node IDs}

\item{parentNodeIDs}{A vector of parent node IDs}

\item{leftChildIDs}{A vector of left-child node IDs}

\item{rightChildIDs}{A vector of right-child node IDs}
}
\description{
Make a btree, data.table object from one of two constructors:

\itemize{
 \item{A vector of node IDs and parent node IDs} or
 \item{A vector of node IDs, left-child IDs, and right-child IDs}
}
}
\details{
Returns a data.table object with fields {NodeID, ParentNodeID, LeftChildNodeID, RightChildNodeID}
}
\examples{
library(data.table)

# constructor 1
make_btree(nodeIDs=c(1), parentNodeIDs=c(NA_integer_))
make_btree(nodeIDs=c(1,2,3), parentNodeIDs=c(NA,1,1))
make_btree(nodeIDs=c(1,2,3,4,5,6,7), parentNodeIDs=c(NA,1,1,2,2,3,3))

# constructor 2
make_btree(nodeIDs=c(1), leftChildIDs=c(NA_integer_), rightChildIDs=c(NA_integer_))
make_btree(nodeIDs=c(1), leftChildIDs=c(2), rightChildIDs=c(3))
make_btree(nodeIDs=c(1,2,3), leftChildIDs=c(2,NA,NA), rightChildIDs=c(3,NA,NA))
}

