% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dtree.R
\name{make_dtree}
\alias{make_dtree}
\title{Make dtree}
\usage{
make_dtree(nodeIds, leftChildIds, rightChildIds, splitVars, splitVals,
  catVars = NULL)
}
\arguments{
\item{nodeIds}{A vector of node Ids}

\item{leftChildIds}{A vector of left-child node Ids}

\item{rightChildIds}{A vector of right-child node Ids}

\item{parentNodeIds}{A vector of parent node Ids}
}
\description{
Make a dtree from one of two constructors:

\itemize{
 \item{A vector of node Ids and parent node Ids} or
 \item{A vector of node Ids, left-child Ids, and right-child Ids}
}
}
\details{
Returns a dtree object
}
\examples{
library(data.table)

# constructor 1
make_btree(nodeIds=c(1), parentNodeIds=c(NA_integer_))
make_btree(nodeIds=c(1,2,3), parentNodeIds=c(NA,1,1))
make_btree(nodeIds=c(1,2,3,4,5,6,7), parentNodeIds=c(NA,1,1,2,2,3,3))

# constructor 2
make_btree(nodeIds=c(1), leftChildIds=c(NA_integer_), rightChildIds=c(NA_integer_))
make_btree(nodeIds=c(1), leftChildIds=c(2), rightChildIds=c(3))
make_btree(nodeIds=c(1,2,3), leftChildIds=c(2,NA,NA), rightChildIds=c(3,NA,NA))
}
