% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_perfect_btree.R
\name{make_perfect_btree}
\alias{make_perfect_btree}
\title{Make Perfect btree}
\usage{
make_perfect_btree(depth)
}
\arguments{
\item{depth}{Depth of the btree}
}
\description{
Generate a perfect btree
}
\details{
Returns a btree object that is a perfect binary tree
}
\examples{
library(data.table)
make_perfect_btree(2)
}
\references{
\url{https://en.wikipedia.org/wiki/Binary_tree#Types_of_binary_trees}
}
