% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_depths.R
\name{node_depths}
\alias{node_depths}
\title{Node Depths}
\usage{
node_depths(btree, nodeIds = NULL)
}
\arguments{
\item{btree}{A btree object}

\item{nodeIds}{Optionally a vector of node Ids corresponding to nodes in btree}
}
\description{
Given a btree, determine the depth of each node
}
\details{
Returns a vector of node depths, with root node depth = 0
}
\examples{
library(data.table)

btree1 <- make_btree(nodeIds=c(1,2,3,4,5,6,7), parentNodeIds=c(NA,1,1,2,2,3,3))
node_depths(btree1)
}
