% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_btree.R
\name{plot_btree}
\alias{plot_btree}
\title{Plot btree}
\usage{
plot_btree(btree, labelCol = "NodeID")
}
\arguments{
\item{btree}{btree to plot}

\item{labelCol}{name of the column whose values should be plotted as node labels/info}
}
\description{
Plot a btree object using ggplot2
}
\details{
Returns a ggplot plot of a btree
}
\examples{
library(data.table)
library(ggplot2)
mytree <- make_perfect_btree(2)
plot_btree(mytree)
}

