% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_btree.R
\name{valid_btree}
\alias{valid_btree}
\title{Valid btree}
\usage{
valid_btree(btree, verbose = FALSE)
}
\arguments{
\item{btree}{An object of class(btree)}
}
\description{
Check if an object of type btree is valid
}
\details{
Returns TRUE if
}
\examples{
library(data.table)
mytree <- make_btree(nodeIds=c(1,2,3,4,5,6,7), parentNodeIds=c(NA,1,1,2,2,3,3))
sub_btree(mytree, nodeId=3)
sub_btree(mytree, depth=1)

}
