% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtree_from_randomForest.R
\name{dtree_from_randomForest}
\alias{dtree_from_randomForest}
\title{Extract a dtree from a randomForest object}
\usage{
dtree_from_randomForest(rf, kthTree)
}
\arguments{
\item{rf}{A randomForest object}

\item{kthTree}{Integer specifying which tree to extract from rf}
}
\description{
Create a dtree object from a tree from a randomForest object
}
\details{
Returns a dtree object
}
\examples{
library(data.table)
mytree <- data.table(NodeId=c(1, 2, 3), ParentNodeId=c(NA, 1, 1))
btree_paths(mytree)
}
