% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dtree.R
\name{make_dtree}
\alias{make_dtree}
\title{Make dtree}
\usage{
make_dtree(nodeIds, leftChildIds, rightChildIds, splitVars = NULL,
  splitVals = NULL, data = NULL)
}
\arguments{
\item{nodeIds}{A vector of node Ids}

\item{leftChildIds}{A vector of left-child node Ids}

\item{rightChildIds}{A vector of right-child node Ids}

\item{parentNodeIds}{A vector of parent node Ids}
}
\description{
Make a dtree from one of two constructors:

\itemize{
 \item{A vector of node Ids and parent node Ids} or
 \item{A vector of node Ids, left-child Ids, and right-child Ids}
}
}
\details{
Returns a dtree object
}
\examples{
library(data.table)

make_dtree(nodeIds=c(1), leftChildIds=c(NA_integer_), rightChildIds=c(NA_integer_))
make_dtree(nodeIds=c(1), leftChildIds=c(2), rightChildIds=c(3), splitVars=c("foo"), splitVals=c(3.5))
make_dtree(nodeIds=c(1,2,3), leftChildIds=c(2,NA,NA), rightChildIds=c(3,NA,NA), splitVars=c("foo",NA,NA), splitVals=c(3.5,NA,NA))
}
