% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_btree.R
\name{sub_btree}
\alias{sub_btree}
\title{Sub btree}
\usage{
sub_btree(btree, nodeID)
}
\arguments{
\item{btree}{A data.table object that is a btree}

\item{nodeID}{ID of the node which should be the root node in the sub-btree}
}
\description{
Extract a sub-btree from an existing btree
}
\details{
Returns a btree, data.table
}
\examples{
library(data.table)
mytree <- make_btree(nodeIDs=c(1,2,3,4,5,6,7), parentNodeIDs=c(NA,1,1,2,2,3,3))
sub_btree(mytree, nodeID=3)
}

